Added function for doing surfrevs in 2D (i.e. making circles)
The generateSurfrev2D() function (presently in Crate.cpp, might be moved later) generates a 2-D cross-section of a "slice" of a spherical shell. NUM_SLICES can be changed to adjust the level of detail (since this is, of course, just an approximation of a circle).
(generateSurfrev3D(), currently a placeholder, will take the output of this and surfrev it into a 3-D shape.)

Also, I copied Box's init() and draw() functions into Layer, since we'll be adapting them to draw the new mesh.